rename constant package name
