+++ basic/internal/constantsample/goconst.go
@@ -1,4 +1,4 @@
-package basicsyntax
+package constantsample
 
 import (
 	"fmt"
