--- basic/internal/pkg1/fileutil.go
@@ -1,7 +1,0 @@
-package pkg1
-
-import "fmt"
-
-func init() {
-	fmt.Println("pkg1.fileutil init here...")
-}
